Modularize messages handler
