+++ pasha-bot/handlers/messages/api_client.py
@@ -1,0 +1,38 @@
+
+import logging
+import google.generativeai as genai
+from config import GEMINI_API_KEY
+
+# Load the Gemini API key
+genai.configure(api_key=GEMINI_API_KEY)
+
+def get_gemini_summary(prompt: str) -> str:
+    # Define the model configuration
+    generation_config = {
+        "temperature": 0,
+        "top_p": 0.95,
+        "top_k": 64,
+        "max_output_tokens": 8192,
+        "response_mime_type": "text/plain",
+    }
+
+    try:
+        # Log before making the API request
+        logging.info(f"Sending prompt to Gemini API: {prompt}")
+
+        # Create the model
+        model = genai.GenerativeModel(
+            model_name="gemini-1.5-flash",
+            generation_config=generation_config,
+        )
+
+        # Start a conversation and send the prompt
+        chat_session = model.start_chat()
+        response = chat_session.send_message(f"{prompt}")
+
+        logging.info(f"Received response from Gemini API: {response}")
+
+        return response.text if response else "No response from the Gemini API."
+    except Exception as e:
+        logging.error(f"Error in Gemini API call: {str(e)}")
+        return "An error occurred while calling the Gemini API."
